+++ packages/dialtone-vue3/components/rich_text_editor/extensions/image/index.js
@@ -1,0 +1,3 @@
+import { ConfigurableImage } from './image';
+
+export default ConfigurableImage;