+++ packages/dialtone-vue3/recipes/conversation_view/message_input/message_input.vue
@@ -389,7 +389,7 @@
           gap="300"
         >
           <!-- @slot Slot for sms count -->
-          <div class="d-d-flex d-ai-center">
+          <div class="d-recipe-message-input__sms-count">
             <slot name="smsCount" />
           </div>